Add DisableAuthentication server option.
This commit allows one to put `DisableAuthentication yes` into an
sshd_config.

When this option is given, sshd will behave as follows:
 * If sshd is running as an ordinary user, then this user can login
   without authentication.
 * If sshd is running as root, then all ordinary users can login without
   authentication. If `PermitRootLogin yes` is also in the config, then
   root can login without authentication.

It is recommended to only use this option where authentication is
achieved by other means and ssh is only used for its other features,
such as port forwarding.

Example: Run this on an unprivileged port inside a docker container that
requires authenticated kubectl port-forward to reach at all.